Hi. I'm actually learning the basics of phyton i need someone to help :)
I am tasked by my school to create a code that allows the user to find out the resistance of 3 resistors be it in series or parallel. When i key in series. it gives me the correct output but when i key in parallel, it gives me the calculation of series. what am i doing wrong